refactor: implement conftest.py for database tests
